Split out the response methods.
The Lambda class had a lot of different methods in it and it was
difficult to test what happens under different environments because the
environment is passed in as an environment variable.

I've moved the response and cookies methods out and added another test
class testing for the allowed origin response.

I've not added any tests for the cookies method as this is mostly using
AWS SDK to generate things and as the response is encrypted anyway,
there's not much we can check.