Add the source ip to the policy and change the expiry date.
I've also added in a test for the cookie policy as it's not encrypted,
it's only base64 encoded.
I'm using a TimeUtils trait so I can always pass in the same instant to
the cookie creation class to I can test that the timeouts are right.
I'm also checking the source ip, the resource url and the key pair id.