Remove the requestContext from the case class.
We were getting the source IP from the incoming requestContext object and then adding this to the restrictions in the Cloudfront policy. The problem is that the source IP is the IP from Cloudfront and not the client IP so we can't use this.

There are ways around this which I've put in [this ticket](https://national-archives.atlassian.net/browse/TDR-1397) in the backlog. For now, I'll leave the IP restrictions open.